% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_generateReport.R
\name{generateReport}
\alias{generateReport}
\title{generateReport}
\usage{
generateReport(.inputDatasetUrl, .reportOutputUrl, .fileName)
}
\arguments{
\item{.inputDatasetUrl}{direct link to the .RDS file that has the check report output}

\item{.reportOutputUrl}{link to the folder where the report will be stored}

\item{.fileName}{name of the output file}
}
\description{
generateReport
}
