% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_printCriticalCheck1.R
\name{printCriticalCheck}
\alias{printCriticalCheck}
\title{printCriticalCheck}
\usage{
printCriticalCheck(.criticalCheckId, .criticalCheckResults)
}
\arguments{
\item{.criticalCheckId}{The id of the check to be printed (i.e. criticalCheck1 - criticalCheck9)}

\item{.criticalCheckResults}{The listing output to be printed}
}
\description{
printCriticalCheck
}
