% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_printCriticalChecks.R
\name{printCriticalChecks}
\alias{printCriticalChecks}
\title{printCriticalChecks}
\usage{
printCriticalChecks(.criticalChecksToPrint)
}
\arguments{
\item{.criticalChecksToPrint}{An r dataframe with the critical checks to print}
}
\description{
printCriticalChecks
}
